Fix compilation errors in translated scala files

+++ web/cairo/app/controllers/logged/modules/general/Camion.scala
@@ -19,7 +19,7 @@ case class CamionData(
               esSemi: Boolean,
               patente: String,
               patentesemi: String,
-              tara: Long,
+              tara: Int,
               transId: Int,
               chofId: Int,
               descrip: String
